[HW12] Reorganized storage package structure
