+++ hw12_13_14_15_calendar/internal/storage/abstract/storage.go
@@ -1,0 +1,9 @@
+package abstractstorage
+
+import "context"
+
+type Storage interface {
+	Events() EventRepository
+	Connect(ctx context.Context) error
+	Close() error
+}
