+++ hw12_13_14_15_calendar/internal/storage/sql/storage.go
@@ -3,11 +3,11 @@ package sqlstorage
 import (
 	"context"
 
+	abstractstorage "github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage/abstract"
+
 	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/config"
 
 	"github.com/jmoiron/sqlx"
-
-	"github.com/pls87/OtusGolang_homework/hw12_13_14_15_calendar/internal/storage"
 
 	_ "github.com/lib/pq"
 )
@@ -25,7 +25,7 @@ func New(cfg config.StorageConf) *SQLStorage {
 	}
 }
 
-func (s *SQLStorage) Events() storage.EventRepository {
+func (s *SQLStorage) Events() abstractstorage.EventRepository {
 	return &s.events
 }
